Move out fixtures to conftest
