+++ test/test_ocsp.py
@@ -1,91 +1,13 @@
 import datetime
 
 import pytest
-from cryptography import x509
 from cryptography.hazmat.primitives import hashes, serialization
 from cryptography.hazmat.primitives._serialization import Encoding
-from cryptography.hazmat.primitives.asymmetric import rsa
-from cryptography.x509 import NameOID, ocsp
+from cryptography.x509 import ocsp
 
 from pki_tools.exceptions import ExtensionMissing, Revoked, OcspFetchFailure
 from pki_tools.ocsp import check_revoked_pem
-
-TEST_ACCESS_DESCRIPTION = "test-url"
-
-
-@pytest.fixture()
-def mocked_requests_get(mocker):
-    return mocker.patch("requests.get")
-
-
-@pytest.fixture()
-def key_pair():
-    return rsa.generate_private_key(
-        public_exponent=65537,
-        key_size=2048,
-    )
-
-
-@pytest.fixture()
-def cert(key_pair):
-    return _create_cert(key_pair)
-
-
-def _create_cert(key_pair, add_aia_extension=True):
-    subject = issuer = x509.Name(
-        [
-            x509.NameAttribute(NameOID.COUNTRY_NAME, "US"),
-            x509.NameAttribute(NameOID.STATE_OR_PROVINCE_NAME, "California"),
-            x509.NameAttribute(NameOID.LOCALITY_NAME, "San Francisco"),
-            x509.NameAttribute(NameOID.ORGANIZATION_NAME, "My Company"),
-            x509.NameAttribute(NameOID.COMMON_NAME, "mysite.com"),
-        ]
-    )
-
-    cert_builder = (
-        x509.CertificateBuilder()
-        .subject_name(
-            subject,
-        )
-        .issuer_name(
-            issuer,
-        )
-        .serial_number(
-            x509.random_serial_number(),
-        )
-        .public_key(
-            key_pair.public_key(),
-        )
-        .not_valid_before(
-            datetime.datetime.utcnow(),
-        )
-        .not_valid_after(
-            datetime.datetime.utcnow() + datetime.timedelta(days=10),
-        )
-        .add_extension(
-            x509.SubjectAlternativeName([x509.DNSName("localhost")]),
-            critical=False,
-        )
-    )
-
-    if add_aia_extension:
-        cert_builder = cert_builder.add_extension(
-            x509.AuthorityInformationAccess(
-                [
-                    x509.AccessDescription(
-                        access_method=x509.AuthorityInformationAccessOID.OCSP,
-                        access_location=x509.UniformResourceIdentifier(
-                            value="test_server",
-                        ),
-                    )
-                ]
-            ),
-            critical=False,
-        )
-
-    cert = cert_builder.sign(key_pair, hashes.SHA256())
-
-    return cert
+from conftest import _create_cert
 
 
 def _create_mocked_ocsp_response(
